Fixes issue where angle of spread weapons was constant and too wide. Minor cleanup, makes certain things be excluded from compile via DEBUG flag

+++ exam_project/utilities/components.hpp
@@ -34,8 +34,8 @@ register_component(EnemyData)
 
 struct Health
 {
-	float max;
-	float curr;
+	int max;
+	int curr;
 	float elapsed;
 	float grace_period;
 };
